+++ R/label.R
@@ -121,12 +121,12 @@ catto_label <- function(train,
 #' @export
 catto_label.data.frame <- function(train,
                                    ...,
-                                   test,
+                                   test = NULL,
                                    ordering = "increasing",
                                    verbose = TRUE,
                                    seed = 4444) {
   validate_col_types(train)
-  test_also <- !missing(test)
+  test_also <- !is.null(test)
   if (test_also) check_train_test(train, test)
 
   cats <- pick_cols(train, deparse(substitute(train)), ...)